Refactor runInTerminalTool to simplify prepare call
