+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/tools/runInTerminalTool.ts
@@ -4,47 +4,44 @@
  *--------------------------------------------------------------------------------------------*/
 
 import type { IMarker as IXtermMarker } from '@xterm/xterm';
+import { asArray } from '../../../../../../base/common/arrays.js';
 import { timeout } from '../../../../../../base/common/async.js';
 import { CancellationToken } from '../../../../../../base/common/cancellation.js';
+import { Codicon } from '../../../../../../base/common/codicons.js';
 import { CancellationError } from '../../../../../../base/common/errors.js';
 import { MarkdownString, type IMarkdownString } from '../../../../../../base/common/htmlContent.js';
 import { Disposable, DisposableStore } from '../../../../../../base/common/lifecycle.js';
+import { basename } from '../../../../../../base/common/path.js';
 import { OperatingSystem, OS } from '../../../../../../base/common/platform.js';
 import { count } from '../../../../../../base/common/strings.js';
-import type { URI } from '../../../../../../base/common/uri.js';
+import type { SingleOrMany } from '../../../../../../base/common/types.js';
 import { generateUuid } from '../../../../../../base/common/uuid.js';
 import { localize } from '../../../../../../nls.js';
+import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
 import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
 import { IStorageService, StorageScope, StorageTarget } from '../../../../../../platform/storage/common/storage.js';
 import { TerminalCapability } from '../../../../../../platform/terminal/common/capabilities/capabilities.js';
 import { ITerminalLogService } from '../../../../../../platform/terminal/common/terminal.js';
-import { IWorkspaceContextService } from '../../../../../../platform/workspace/common/workspace.js';
 import { IRemoteAgentService } from '../../../../../services/remote/common/remoteAgentService.js';
 import { IChatService, type IChatTerminalToolInvocationData } from '../../../../chat/common/chatService.js';
 import { CountTokensCallback, ILanguageModelToolsService, IPreparedToolInvocation, IToolData, IToolImpl, IToolInvocation, IToolInvocationPreparationContext, IToolResult, ToolDataSource, ToolProgress, type IToolConfirmationMessages, type ToolConfirmationAction } from '../../../../chat/common/languageModelToolsService.js';
-import { Separator } from '../../../../../../base/common/actions.js';
 import { ITerminalService, type ITerminalInstance } from '../../../../terminal/browser/terminal.js';
 import type { XtermTerminal } from '../../../../terminal/browser/xterm/xtermTerminal.js';
 import { ITerminalProfileResolverService } from '../../../../terminal/common/terminal.js';
+import { TerminalChatAgentToolsSettingId } from '../../common/terminalChatAgentToolsConfiguration.js';
 import { getRecommendedToolsOverRunInTerminal } from '../alternativeRecommendation.js';
 import { getOutput } from '../bufferOutputPolling.js';
-import { CommandLineAutoApprover, type IAutoApproveRule, type ICommandApprovalResult, type ICommandApprovalResultWithReason } from '../commandLineAutoApprover.js';
+import { CommandLineAutoApprover, type IAutoApproveRule, type ICommandApprovalResult } from '../commandLineAutoApprover.js';
+import { CommandSimplifier } from '../commandSimplifier.js';
 import { BasicExecuteStrategy } from '../executeStrategy/basicExecuteStrategy.js';
 import type { ITerminalExecuteStrategy } from '../executeStrategy/executeStrategy.js';
 import { NoneExecuteStrategy } from '../executeStrategy/noneExecuteStrategy.js';
 import { RichExecuteStrategy } from '../executeStrategy/richExecuteStrategy.js';
-import { isPowerShell } from '../runInTerminalHelpers.js';
+import { OutputMonitor } from '../outputMonitor.js';
+import { generateAutoApproveActions, isPowerShell } from '../runInTerminalHelpers.js';
+import { RunInTerminalToolTelemetry } from '../runInTerminalToolTelemetry.js';
 import { splitCommandLineIntoSubCommands } from '../subCommands.js';
 import { ShellIntegrationQuality, ToolTerminalCreator, type IToolTerminal } from '../toolTerminalCreator.js';
-import { Codicon } from '../../../../../../base/common/codicons.js';
-import { OutputMonitor } from '../outputMonitor.js';
-import type { TerminalNewAutoApproveButtonData } from '../../../../chat/browser/chatContentParts/toolInvocationParts/chatTerminalToolConfirmationSubPart.js';
-import { basename } from '../../../../../../base/common/path.js';
-import type { SingleOrMany } from '../../../../../../base/common/types.js';
-import { asArray } from '../../../../../../base/common/arrays.js';
-import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
-import { TerminalChatAgentToolsSettingId } from '../../common/terminalChatAgentToolsConfiguration.js';
-import { RunInTerminalToolTelemetry } from '../runInTerminalToolTelemetry.js';
 
 const TERMINAL_SESSION_STORAGE_KEY = 'chat.terminalSessions';
 
@@ -141,6 +138,7 @@ const promptInjectionWarningCommandsLowerPwshOnly = [
 export class RunInTerminalTool extends Disposable implements IToolImpl {
 
 	private readonly _terminalToolCreator: ToolTerminalCreator;
+	private readonly _commandSimplifier: CommandSimplifier;
 	private readonly _telemetry: RunInTerminalToolTelemetry;
 	protected readonly _commandLineAutoApprover: CommandLineAutoApprover;
 	protected readonly _sessionTerminalAssociations: Map<string, IToolTerminal> = new Map();
@@ -167,14 +165,15 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 		@ITerminalService private readonly _terminalService: ITerminalService,
 		@IRemoteAgentService private readonly _remoteAgentService: IRemoteAgentService,
 		@IChatService private readonly _chatService: IChatService,
-		@IWorkspaceContextService private readonly _workspaceContextService: IWorkspaceContextService,
 	) {
 		super();
 
+		this._osBackend = this._remoteAgentService.getEnvironment().then(remoteEnv => remoteEnv?.os ?? OS);
+
 		this._terminalToolCreator = _instantiationService.createInstance(ToolTerminalCreator);
+		this._commandSimplifier = _instantiationService.createInstance(CommandSimplifier, this._osBackend);
 		this._telemetry = _instantiationService.createInstance(RunInTerminalToolTelemetry);
 		this._commandLineAutoApprover = this._register(_instantiationService.createInstance(CommandLineAutoApprover));
-		this._osBackend = this._remoteAgentService.getEnvironment().then(remoteEnv => remoteEnv?.os ?? OS);
 
 		// Restore terminal associations from storage
 		this._restoreTerminalAssociations();
@@ -208,7 +207,7 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 		const instance = context.chatSessionId ? this._sessionTerminalAssociations.get(context.chatSessionId)?.instance : undefined;
 		const terminalToolSessionId = generateUuid();
 
-		let toolEditedCommand: string | undefined = await this._rewriteCommandIfNeeded(args, instance, shell);
+		let toolEditedCommand: string | undefined = await this._commandSimplifier.rewriteIfNeeded(args, instance, shell);
 		if (toolEditedCommand === args.command) {
 			toolEditedCommand = undefined;
 		}
@@ -262,59 +261,18 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 				}
 			}
 
-			function formatRuleLinks(result: SingleOrMany<{ result: ICommandApprovalResult; rule?: IAutoApproveRule; reason: string }>): string {
-				return asArray(result).map(e => {
-					return `[\`${e.rule!.sourceText}\`](settings_${e.rule!.sourceTarget} "${localize('ruleTooltip', 'View rule in settings')}")`;
-				}).join(', ');
-			}
-
-			const config = this._configurationService.inspect<boolean | Record<string, boolean>>('chat.tools.autoApprove');
-			const isGlobalAutoApproved = config?.value ?? config.defaultValue;
-			if (isGlobalAutoApproved) {
-				autoApproveInfo = new MarkdownString(`_${localize('autoApprove.global', 'Auto approved by setting {0}', `[\`chat.tools.autoApprove\`](settings_global "${localize('ruleTooltip.global', 'View settings')}")`)}_`);
-			} else if (isAutoApproved) {
-				switch (autoApproveReason) {
-					case 'commandLine': {
-						if (commandLineResult.rule) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApprove.rule', 'Auto approved by rule {0}', formatRuleLinks(commandLineResult))}_`);
-						}
-						break;
-					}
-					case 'subCommand': {
-						const uniqueRules = Array.from(new Set(subCommandResults));
-						if (uniqueRules.length === 1) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApprove.rule', 'Auto approved by rule {0}', formatRuleLinks(uniqueRules))}_`);
-						} else if (uniqueRules.length > 1) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApprove.rules', 'Auto approved by rules {0}', formatRuleLinks(uniqueRules))}_`);
-						}
-						break;
-					}
-				}
-			} else if (isDenied) {
-				switch (autoApproveReason) {
-					case 'commandLine': {
-						if (commandLineResult.rule) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApproveDenied.rule', 'Auto approval denied by rule {0}', formatRuleLinks(commandLineResult))}_`);
-						}
-						break;
-					}
-					case 'subCommand': {
-						const deniedRules = subCommandResults.filter(e => e.result === 'denied');
-						const uniqueRules = Array.from(new Set(deniedRules));
-						if (uniqueRules.length === 1) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApproveDenied.rule', 'Auto approval denied by rule {0}', formatRuleLinks(uniqueRules))}_`);
-						} else if (uniqueRules.length > 1) {
-							autoApproveInfo = new MarkdownString(`_${localize('autoApproveDenied.rules', 'Auto approval denied by rules {0}', formatRuleLinks(uniqueRules))}_`);
-						}
-						break;
-					}
-				}
-			}
-
 			// Log detailed auto approval reasoning
 			for (const reason of autoApproveReasons) {
 				this._logService.info(`- ${reason}`);
 			}
+
+			autoApproveInfo = this._createAutoApproveInfo(
+				isAutoApproved,
+				isDenied,
+				autoApproveReason,
+				subCommandResults,
+				commandLineResult,
+			);
 
 			// TODO: Move this higher, prevent unnecessary work
 			const isAutoApproveEnabled = this._configurationService.getValue(TerminalChatAgentToolsSettingId.EnableAutoApprove) === 'on';
@@ -343,7 +301,7 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 
 			let customActions: ToolConfirmationAction[] | undefined;
 			if (!isAutoApproved && isAutoApproveEnabled) {
-				customActions = this._generateAutoApproveActions(actualCommand, subCommands, { subCommandResults, commandLineResult });
+				customActions = generateAutoApproveActions(actualCommand, subCommands, { subCommandResults, commandLineResult });
 			}
 
 			let shellType = basename(shell, '.exe');
@@ -607,6 +565,8 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 		}
 	}
 
+	// #region Terminal init
+
 	private async _initBackgroundTerminal(chatSessionId: string, termId: string, token: CancellationToken): Promise<IToolTerminal> {
 		this._logService.debug(`RunInTerminalTool: Creating background terminal with ID=${termId}`);
 		const toolTerminal = await this._terminalToolCreator.createTerminal(token);
@@ -636,60 +596,9 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 		return toolTerminal;
 	}
 
-	protected async _rewriteCommandIfNeeded(args: IRunInTerminalInputParams, instance: Pick<ITerminalInstance, 'getCwdResource'> | undefined, shell: string): Promise<string> {
-		const commandLine = args.command;
-		const os = await this._osBackend;
+	// #endregion
 
-		// Re-write the command if it starts with `cd <dir> && <suffix>` or `cd <dir>; <suffix>`
-		// to just `<suffix>` if the directory matches the current terminal's cwd. This simplifies
-		// the result in the chat by removing redundancies that some models like to add.
-		const isPwsh = isPowerShell(shell, os);
-		const cdPrefixMatch = commandLine.match(
-			isPwsh
-				? /^(?:cd(?: \/d)?|Set-Location(?: -Path)?) (?<dir>[^\s]+) ?(?:&&|;)\s+(?<suffix>.+)$/i
-				: /^cd (?<dir>[^\s]+) &&\s+(?<suffix>.+)$/
-		);
-		const cdDir = cdPrefixMatch?.groups?.dir;
-		const cdSuffix = cdPrefixMatch?.groups?.suffix;
-		if (cdDir && cdSuffix) {
-			let cwd: URI | undefined;
-
-			// Get the current session terminal's cwd
-			if (instance) {
-				cwd = await instance.getCwdResource();
-			}
-
-			// If a terminal is not available, use the workspace root
-			if (!cwd) {
-				const workspaceFolders = this._workspaceContextService.getWorkspace().folders;
-				if (workspaceFolders.length === 1) {
-					cwd = workspaceFolders[0].uri;
-				}
-			}
-
-			// Re-write the command if it matches the cwd
-			if (cwd) {
-				// Remove any surrounding quotes
-				let cdDirPath = cdDir;
-				if (cdDirPath.startsWith('"') && cdDirPath.endsWith('"')) {
-					cdDirPath = cdDirPath.slice(1, -1);
-				}
-				// Normalize trailing slashes
-				cdDirPath = cdDirPath.replace(/(?:[\\\/])$/, '');
-				let cwdFsPath = cwd.fsPath.replace(/(?:[\\\/])$/, '');
-				// Case-insensitive comparison on Windows
-				if (os === OperatingSystem.Windows) {
-					cdDirPath = cdDirPath.toLowerCase();
-					cwdFsPath = cwdFsPath.toLowerCase();
-				}
-				if (cdDirPath === cwdFsPath) {
-					return cdSuffix;
-				}
-			}
-		}
-
-		return commandLine;
-	}
+	// #region Session management
 
 	private _restoreTerminalAssociations(): void {
 		const storedAssociations = this._storageService.get(TERMINAL_SESSION_STORAGE_KEY, StorageScope.WORKSPACE, '{}');
@@ -795,98 +704,72 @@ export class RunInTerminalTool extends Disposable implements IToolImpl {
 		}
 	}
 
-	private _generateAutoApproveActions(commandLine: string, subCommands: string[], autoApproveResult: { subCommandResults: ICommandApprovalResultWithReason[]; commandLineResult: ICommandApprovalResultWithReason }): ToolConfirmationAction[] {
-		const actions: ToolConfirmationAction[] = [];
+	// #endregion
 
-		// We shouldn't offer configuring rules for commands that are explicitly denied since it
-		// wouldn't get auto approved with a new rule
-		const canCreateAutoApproval = autoApproveResult.subCommandResults.some(e => e.result !== 'denied') || autoApproveResult.commandLineResult.result === 'denied';
-		if (canCreateAutoApproval) {
-			const unapprovedSubCommands = subCommands.filter((_, index) => {
-				return autoApproveResult.subCommandResults[index].result !== 'approved';
-			});
+	// #region Auto approve
 
-			// For each unapproved sub-command (within the overall command line), decide whether to
-			// suggest just the commnad or sub-command (with that sub-command line) to always allow.
-			const commandsWithSubcommands = new Set(['git', 'npm', 'yarn', 'docker', 'kubectl', 'cargo', 'dotnet', 'mvn', 'gradle']);
-			const commandsWithSubSubCommands = new Set(['npm run', 'yarn run']);
-			const subCommandsToSuggest = Array.from(new Set(unapprovedSubCommands.map(command => {
-				const parts = command.trim().split(/\s+/);
-				const baseCommand = parts[0].toLowerCase();
-				const baseSubCommand = parts.length > 1 ? `${parts[0]} ${parts[1]}`.toLowerCase() : '';
+	private _createAutoApproveInfo(
+		isAutoApproved: boolean,
+		isDenied: boolean,
+		autoApproveReason: 'subCommand' | 'commandLine' | undefined,
+		subCommandResults: ReturnType<CommandLineAutoApprover['isCommandAutoApproved']>[],
+		commandLineResult: ReturnType<CommandLineAutoApprover['isCommandLineAutoApproved']>,
+	): MarkdownString | undefined {
+		function formatRuleLinks(result: SingleOrMany<{ result: ICommandApprovalResult; rule?: IAutoApproveRule; reason: string }>): string {
+			return asArray(result).map(e => {
+				return `[\`${e.rule!.sourceText}\`](settings_${e.rule!.sourceTarget} "${localize('ruleTooltip', 'View rule in settings')}")`;
+			}).join(', ');
+		}
 
-				if (commandsWithSubSubCommands.has(baseSubCommand) && parts.length >= 3 && !parts[2].startsWith('-')) {
-					return `${parts[0]} ${parts[1]} ${parts[2]}`;
-				} else if (commandsWithSubcommands.has(baseCommand) && parts.length >= 2 && !parts[1].startsWith('-')) {
-					return `${parts[0]} ${parts[1]}`;
-				} else {
-					return parts[0];
+		const config = this._configurationService.inspect<boolean | Record<string, boolean>>('chat.tools.autoApprove');
+		const isGlobalAutoApproved = config?.value ?? config.defaultValue;
+		if (isGlobalAutoApproved) {
+			return new MarkdownString(`_${localize('autoApprove.global', 'Auto approved by setting {0}', `[\`chat.tools.autoApprove\`](settings_global "${localize('ruleTooltip.global', 'View settings')}")`)}_`);
+		}
+
+		if (isAutoApproved) {
+			switch (autoApproveReason) {
+				case 'commandLine': {
+					if (commandLineResult.rule) {
+						return new MarkdownString(`_${localize('autoApprove.rule', 'Auto approved by rule {0}', formatRuleLinks(commandLineResult))}_`);
+					}
+					break;
 				}
-			})));
-
-			if (subCommandsToSuggest.length > 0) {
-				let subCommandLabel: string;
-				let subCommandTooltip: string;
-				if (subCommandsToSuggest.length === 1) {
-					subCommandLabel = localize('autoApprove.baseCommandSingle', 'Always Allow Command: {0}', subCommandsToSuggest[0]);
-					subCommandTooltip = localize('autoApprove.baseCommandSingleTooltip', 'Always allow command starting with `{0}` to run without confirmation', subCommandsToSuggest[0]);
-				} else {
-					const commandSeparated = subCommandsToSuggest.join(', ');
-					subCommandLabel = localize('autoApprove.baseCommand', 'Always Allow Commands: {0}', commandSeparated);
-					subCommandTooltip = localize('autoApprove.baseCommandTooltip', 'Always allow commands starting with `{0}` to run without confirmation', commandSeparated);
+				case 'subCommand': {
+					const uniqueRules = Array.from(new Set(subCommandResults));
+					if (uniqueRules.length === 1) {
+						return new MarkdownString(`_${localize('autoApprove.rule', 'Auto approved by rule {0}', formatRuleLinks(uniqueRules))}_`);
+					} else if (uniqueRules.length > 1) {
+						return new MarkdownString(`_${localize('autoApprove.rules', 'Auto approved by rules {0}', formatRuleLinks(uniqueRules))}_`);
+					}
+					break;
 				}
-
-				actions.push({
-					label: subCommandLabel,
-					tooltip: subCommandTooltip,
-					data: {
-						type: 'newRule',
-						rule: subCommandsToSuggest.map(key => ({
-							key,
-							value: true
-						}))
-					} satisfies TerminalNewAutoApproveButtonData
-				});
 			}
-
-			// Allow exact command line, don't do this if it's just the first sub-command's first
-			// word
-			const firstSubcommandFirstWord = unapprovedSubCommands.length > 0 ? unapprovedSubCommands[0].split(' ')[0] : '';
-			if (firstSubcommandFirstWord !== commandLine) {
-				const truncatedCommandLine = commandLine.length > 40 ? commandLine.substring(0, 40) + '\u2026' : commandLine;
-				actions.push({
-					// Add an extra & since it's treated as a mnemonic
-					label: localize('autoApprove.exactCommand', 'Always Allow Exact Command Line: {0}', truncatedCommandLine.replaceAll('&&', '&&&')),
-					tooltip: localize('autoApprove.exactCommandTooltip', 'Always allow this exact command line to run without confirmation'),
-					data: {
-						type: 'newRule',
-						rule: {
-							key: commandLine,
-							value: {
-								approve: true,
-								matchCommandLine: true
-							}
-						}
-					} satisfies TerminalNewAutoApproveButtonData
-				});
+		} else if (isDenied) {
+			switch (autoApproveReason) {
+				case 'commandLine': {
+					if (commandLineResult.rule) {
+						return new MarkdownString(`_${localize('autoApproveDenied.rule', 'Auto approval denied by rule {0}', formatRuleLinks(commandLineResult))}_`);
+					}
+					break;
+				}
+				case 'subCommand': {
+					const deniedRules = subCommandResults.filter(e => e.result === 'denied');
+					const uniqueRules = Array.from(new Set(deniedRules));
+					if (uniqueRules.length === 1) {
+						return new MarkdownString(`_${localize('autoApproveDenied.rule', 'Auto approval denied by rule {0}', formatRuleLinks(uniqueRules))}_`);
+					} else if (uniqueRules.length > 1) {
+						return new MarkdownString(`_${localize('autoApproveDenied.rules', 'Auto approval denied by rules {0}', formatRuleLinks(uniqueRules))}_`);
+					}
+					break;
+				}
 			}
 		}
 
-		if (actions.length > 0) {
-			actions.push(new Separator());
-		}
-
-		// Always show configure option
-		actions.push({
-			label: localize('autoApprove.configure', 'Configure Auto Approve...'),
-			tooltip: localize('autoApprove.configureTooltip', 'Open settings to configure terminal command auto approval'),
-			data: {
-				type: 'configure'
-			} satisfies TerminalNewAutoApproveButtonData
-		});
-
-		return actions;
+		return undefined;
 	}
+
+	// #endregion
 }
 
 class BackgroundTerminalExecution extends Disposable {
